Metadata for 2023 Quality of Government institute Basic Data - time-series

Generated automatically from spss-version of Quality of Government Basic data using tidymetadata::create_metadata() function. Formatted in tidy-format with row per each value of an indicator. Numeric indicators have a single row per indicator.

Usage

meta_basic_ts_2023

Format

A data frame with 458 rows and five columns

code

Indicator code in data

name

Name of the indicator

value

Indicator value for factor indicator

label

Value label for factor indicators

class

Class of the indicator. Numeric, factor or character

See also